requirements_rpi.txt
RPi.GPIO==0.7.0
rpi_ws281x==5.0.0
